requirements.txt
fastapi==0.72.0
croniter==1.2.0
uvicorn==0.15.0
gunicorn==20.1.0
google-cloud-logging==2.7.0
google-cloud-workflows==1.5.0
pendulum==2.1.2